drivers: mathworks: fix -Werror=int-to-pointer-cast cases
This seems to fail on the Raspberry Pi branch 4.19. It's probable that some
CFLAG has been enabled on the 4.19 kernel that appears in the RPi kernel
but not in Xilinx's.

This change fixes them for the mathworks driver.
The fix is to first cast the value of the int-type to `uintptr_t`. This
matches the size of `void *`. From there-on the value can be cast to
`void *`.

(cherry picked from commit 6360724acb907da53761d13379368b80a03269eb)

+++ drivers/misc/mathworks/mw_stream_channel.c
@@ -879,7 +879,8 @@ static int mw_axidma_alloc(struct mwadma_dev *mwdev, size_t bufferSize)
 
     else {
         dev_info(IP2DEVP(mwdev), "Address of buffer = 0x%p, Length = %u Bytes\n",\
-                (void *)MWDEV_TO_MWIP(mwdev)->dma_info.phys,(unsigned int)bufferSize);
+                (void *)((uintptr_t)MWDEV_TO_MWIP(mwdev)->dma_info.phys),
+		(unsigned int)bufferSize);
         MWDEV_TO_MWIP(mwdev)->dma_info.size = bufferSize;
     }
     return 0;
